backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
pydantic
pydantic-settings
python-dotenv
pandas
numpy
scikit-learn
httpx
pytest
